Fix games failing ROM checksums
Once the cart download finishes, `ioctl_addr` points to the last word as
that's where the last write happened. This value is then stored as the
size of the ROM and passed to the Genesis core.

The core's CPU reads zero for accesses at or above the ROM size. Since
the last word of most games is non-zero, this causes essentially every
game which verifies its checksum to always fail since it will read zero
for the last word instead of its real value, even though that value was
correctly loaded into SDRAM.

Add one word (2 bytes) to the last address to get the correct size and
fix this issue.