Fix red line at the top of the screen
The core's vblank signal turns on and off slightly before the end of the
visible portion of the scanline. This causes pixels to stop being output
before the end of the last visible scanline and start being output
before the end of the last blank scanline.

This latter effect draws a few pixels to the first line on the screen
before the hsync pulse which starts the first visible line occurs. The
remainder of this line renders as junk data from the scaler's
framebuffer, which for whatever reason is usually reddish.

Correct this issue by only sampling vblank at the hsync pulse as that's
when we care whether or not the line will be shown.