ROM and video fixes (#3)
* Fix games failing ROM checksums

Once the cart download finishes, `ioctl_addr` points to the last word as
that's where the last write happened. This value is then stored as the
size of the ROM and passed to the Genesis core.

The core's CPU reads zero for accesses at or above the ROM size. Since
the last word of most games is non-zero, this causes essentially every
game which verifies its checksum to always fail since it will read zero
for the last word instead of its real value, even though that value was
correctly loaded into SDRAM.

Add one word (2 bytes) to the last address to get the correct size and
fix this issue.

* Fix red line at the top of the screen

The core's vblank signal turns on and off slightly before the end of the
visible portion of the scanline. This causes pixels to stop being output
before the end of the last visible scanline and start being output
before the end of the last blank scanline.

This latter effect draws a few pixels to the first line on the screen
before the hsync pulse which starts the first visible line occurs. The
remainder of this line renders as junk data from the scaler's
framebuffer, which for whatever reason is usually reddish.

Correct this issue by only sampling vblank at the hsync pulse as that's
when we care whether or not the line will be shown.